Revert "blah"
This reverts commit 9eafce4d56f85bb988a50b511295187d67d7b9e7.

+++ CSharp/Concentus/Celt/CWRS.cs
@@ -313,7 +313,7 @@ namespace Concentus.Celt
 
         internal static int decode_pulses(int[] _y, int _n, int _k, EntropyCoder _dec)
         {
-            return cwrsi(_n, _k, (uint)_dec.dec_uint(CELT_PVQ_V(_n, _k)), _y);
+            return cwrsi(_n, _k, _dec.dec_uint(CELT_PVQ_V(_n, _k)), _y);
         }
     }
 }
